shopfloor_mobile: rename *.js to *.esm.js
