+++ shopfloor_mobile/static/wms/src/components/batch_picking_detail.esm.js
