+++ shopfloor_mobile/static/wms/src/components/batch_picking_line_detail.esm.js
