+++ shopfloor_mobile/static/wms/src/components/detail/detail_location.esm.js
