+++ shopfloor_mobile/static/wms/src/components/detail/detail_lot.esm.js
