+++ shopfloor_mobile/static/wms/src/components/detail/detail_operation.esm.js
