+++ shopfloor_mobile/static/wms/src/components/detail/detail_package.esm.js
