+++ shopfloor_mobile/static/wms/src/components/detail/detail_picking.esm.js
