+++ shopfloor_mobile/static/wms/src/components/detail/detail_product.esm.js
