+++ shopfloor_mobile/static/wms/src/components/detail/detail_transfer.esm.js
