+++ shopfloor_mobile/static/wms/src/components/forms/form_edit_stock_picking.esm.js
