+++ shopfloor_mobile/static/wms/src/components/manual_select_color.esm.js
