+++ shopfloor_mobile/static/wms/src/components/scenario_picking_detail/picking_summary.esm.js
@@ -4,9 +4,8 @@
  * License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
  */
 
-/* eslint-disable strict */
-/* eslint-disable no-implicit-globals */
-import {PickingDetailListMixin} from "./mixins.js";
+/* eslint-disable sort-imports */
+import {PickingDetailListMixin} from "./mixins.esm.js";
 import {ItemDetailMixin} from "/shopfloor_mobile_base/static/wms/src/components/detail/detail_mixin.js";
 
 Vue.component("picking-summary", {
