+++ shopfloor_mobile/static/wms/src/demo/demo.checkout.esm.js
