+++ shopfloor_mobile/static/wms/src/demo/demo.cluster_picking.esm.js
