+++ shopfloor_mobile/static/wms/src/demo/demo.delivery.esm.js
