+++ shopfloor_mobile/static/wms/src/demo/demo.location_content_transfer.esm.js
