+++ shopfloor_mobile/static/wms/src/demo/demo.scan_anything.esm.js
