+++ shopfloor_mobile/static/wms/src/demo/demo.zone_picking.esm.js
