+++ shopfloor_mobile/static/wms/src/i18n/add_translations_to_registry.esm.js
