+++ shopfloor_mobile/static/wms/src/scenario/checkout_states.esm.js
