+++ shopfloor_mobile/static/wms/src/scenario/cluster_picking.esm.js
