+++ shopfloor_mobile/static/wms/src/scenario/delivery.esm.js
