+++ shopfloor_mobile/static/wms/src/scenario/location_content_transfer.esm.js
