+++ shopfloor_mobile/static/wms/src/scenario/single_pack_transfer.esm.js
