+++ shopfloor_mobile/static/wms/src/scenario/zone_picking.esm.js
