+++ shopfloor_mobile/static/wms/src/screen.esm.js
