+++ shopfloor_mobile/static/wms/src/wms_utils.esm.js
